speaker-sdk-master/examples/requirements.txt
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
grpcio==1.32.0
grpcio-tools==1.32.0
idna==2.10
numpy==1.19.0
prettytable==2.0.0
protobuf==3.14.0
pycparser==2.20
requests==2.25.1
samplerate==0.1.0
six==1.15.0
sounddevice==0.4.0
SoundFile==0.10.3.post1
-e ..
urllib3==1.26.3
wcwidth==0.2.5
